move router and handlers to api folder
main reason is for the router init to be reusable for testing
but also it makes sense in the overall project structure imo
